chore --- finish all TODO list
- support log by slf4j
- add a method for decoding Transaction from String
- add logging method to Plugin so as to write log the lightning daemon
- etc

+++ jightning-api/src/main/java/clightning/apis/response/SimpleInvoice.java
@@ -10,9 +10,8 @@ public class SimpleInvoice {
       @JsonProperty("payment_hash")
       private String paymentHash;
 
-      // TODO: unify the type of timestamp value
       @JsonProperty("expires_at")
-      private long expiresAt;
+      private int expiresAt; // seconds
 
       private String bolt11;
